requirements.txt
python-telegram-bot[job-queue]>=20.7
python-dotenv>=1.0.0
pytz>=2024.1
